Use the entry point to look in several locations for the latest versions of the includes:
The places we look in are based on default github behaviour.  Cloning uses a directory called `omp-stdlib`, while downloading the ZIP uses a directory called `omp-stdlib-master`.  Obviously we can't predict every possible place that a person might save the files, but we can try many common ones.

They are tried in some semblance of a priority.